requirements.txt
Twisted==15.0.0
argparse==1.2.1
autobahn==0.10.2
cffi==0.9.2
characteristic==14.3.0
cryptography==0.8.1
dateutils==0.6.6
enum34==1.0.4
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
python-dateutil==2.4.1
pytz==2015.2
requests==2.6.0
service-identity==14.0.0
six==1.9.0
treq==15.0.0
wsgiref==0.1.2
zope.interface==4.1.2
